# internal/openapi/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "error"
        ],
        "type": "object"
      },
...
      "Ack": {
        "additionalProperties": false,
        "properties": {
          "agent_version": {
            "type": "string"
          },
          "applied_version": {
            "format": "int64",
            "type": "integer"
          },
          "capabilities": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "applied_version",
          "error"
        ],
        "type": "object"
      },
...
      "Control": {
        "additionalProperties": false,
        "properties": {
          "active": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "operation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NodeOperation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "active"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieSession": {
        "in": "cookie",
        "name": "tfp_session",
        "type": "apiKey"
      },
      "nodeBearer": {
        "description": "Separate Agent identity",
        "scheme": "bearer",
        "type": "http"
      },
      "ownerBearer": {
        "description": "Owner resources only, including tokens issued by an administrator",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CNY cents and cumulative byte counts are decimal strings; times are UTC RFC3339. Browser rendering uses Asia/Shanghai. Cookie mutations require X-Requested-With: fetch and matching Origin. Owner Bearer tokens never carry admin privileges. See docs/api-contract.md for transaction, retention and retry semantics.",
    "title": "Traffic Forwarding Panel API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/agent/ack": {
      "post": {
        "operationId": "post_agent_ack",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ack"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [
          {
            "nodeBearer": []
          }
        ],
        "summary": "Versioned configuration acknowledgement",
        "x-role": "node"
      }
    },
    "/agent/config": {
      "get": {
        "operationId": "get_agent_config",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Config"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [
          {
            "nodeBearer": []
          }
        ],
        "summary": "Finite authorized configuration including tunnel secrets",
        "x-role": "node"
      }
    },
    "/agent/control": {
      "post": {
        "operationId": "post_agent_control",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Control"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [
          {
            "nodeBearer": []
          }
        ],
        "summary": "Claim one authorized node operation",
        "x-role": "node"
      }
    },
    "/agent/control/result": {
      "post": {
        "operationId": "post_agent_control_result",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationResult"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [
          {
            "nodeBearer": []
          }
        ],
        "summary": "Report operation completion",
        "x-role": "node"
      }
    },
    "/agent/control/{id}/terminal": {
      "get": {
        "operationId": "get_agent_control_id_terminal",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [
          {
            "nodeBearer": []
          }
        ],
        "summary": "Agent-side terminal WebSocket",
        "x-role": "node"
      }
    },
    "/agent/diagnostics": {
      "post": {
        "operationId": "post_agent_diagnostics",
        "parameters": [
          {
            "description": "Required with Cookie mutations, together with matching Origin; Bearer requests exempt",
            "in": "header",
            "name": "X-Requested-With",
            "schema": {
              "const": "fetch"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticDispatch"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "ownerBearer": []
          }
        ],
        "summary": "Claim one bounded diagnostic task",
        "x-role": "agent"
      }
    },
    "/agent/diagnostics/result": {
      "post": {
        "operationId": "post_agent_diagnostics_result",
        "parameters": [
          {
            "description": "Required with Cookie mutations, together with matching Origin; Bearer requests exempt",
            "in": "header",
            "name": "X-Requested-With",
            "schema": {
              "const": "fetch"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Diagnostic"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "ownerBearer": []
          }
        ],
        "summary": "Submit fenced diagnostic result",
        "x-role": "agent"
      }
    },
    "/agent/leases/retire": {
      "post": {
        "operationId": "post_agent_leases_retire",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetireLease"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [
          {
            "nodeBearer": []
          }
        ],
        "summary": "Retire lease only after final usage is settled",
        "x-role": "node"
      }
    },
    "/agent/looking-glass": {
      "post": {
        "operationId": "post_agent_looking-glass",
        "parameters": [
          {
            "description": "Required with Cookie mutations, together with matching Origin; Bearer requests exempt",
            "in": "header",
            "name": "X-Requested-With",
            "schema": {
              "const": "fetch"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookingGlassDispatch"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "ownerBearer": []
          }
        ],
        "summary": "Claim one pending looking glass request",
        "x-role": "agent"
      }
    },
    "/agent/looking-glass/result": {
      "post": {
        "operationId": "post_agent_looking-glass_result",
        "parameters": [
          {
            "description": "Required with Cookie mutations, together with matching Origin; Bearer requests exempt",
            "in": "header",
            "name": "X-Requested-With",
            "schema": {
              "const": "fetch"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookingGlass"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "ownerBearer": []
          }
        ],
        "summary": "Submit one fenced looking glass result",
        "x-role": "agent"
      }
    },
    "/agent/probe": {
      "post": {
        "operationId": "post_agent_probe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Probe"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [
          {
            "nodeBearer": []
          }
        ],
        "summary": "Submit actual probe observations",
        "x-role": "node"
      }
    },
    "/agent/register": {
      "post": {
        "operationId": "post_agent_register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Registered"
                }
              }
            },
            "description": "Success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Failure; some auth failures are plain text. 409 state conflict/payment_uncertain, 422 unsupported operation."
          }
        },
        "security": [],
        "summary": "Consume one-time enrollment token",
        "x-role": "public"
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}